Move positionOnGrid out of file
